Worked on admin route, get is now roking just dosent reply when no user was found.
